Refactoring test utils into a BaseTestUtilsMixin
The BaseTestUtilsMixin will hold the methods required to create a user, create a room and add users to the room.
Other tests can inherit from the BaseTestUtilsMixin, if they have to create a user or room.
